introduce-response-emitter(move): Hand client-side wait/abort over to ResponseEmitter
Pure physical move per MECH_COMMIT_SPLIT. Cut @staticmethod
_wait_one_response, create_abort_task, _handle_abort_finish_reason,
_coalesce_streaming_chunks from TokenizerManager; paste into
ResponseEmitter (drop @staticmethod, replace
``self: "ResponseEmitter"`` → plain ``self``, strip ``TokenizerManager.``
prefix from sibling-method calls inside moved bodies). Caller prefix
replacement: ``TokenizerManager._wait_one_response(self.response_emitter, ...)``
→ ``self.response_emitter._wait_one_response(...)`` in TM facade;
``tm.create_abort_task(tm.response_emitter, obj)`` →
``tm.response_emitter.create_abort_task(obj)`` in entrypoints.

Refactor chain ID: introduce-response-emitter-move
