+++ python/sglang/srt/entrypoints/anthropic/serving.py
@@ -418,8 +418,7 @@ class AnthropicServing:
                 raw_request,
             ),
             media_type="text/event-stream",
-            background=self.openai_serving_chat.tokenizer_manager.create_abort_task(
-                self.openai_serving_chat.tokenizer_manager.response_emitter,
+            background=self.openai_serving_chat.tokenizer_manager.response_emitter.create_abort_task(
                 adapted_request,
             ),
         )
